# improved performance of Rust implementation of nqueen, added input handling as in other implementation