This may seem counterintuitive, but using static arrays that are twice as large imposes more overhead than allocating just the right amount at runtime.

```
Benchmark 1: ./rust/nqueen
  Time (mean ± σ):      2.455 s ±  0.022 s    [User: 2.414 s, System: 0.038 s]
  Range (min … max):    2.418 s …  2.489 s    10 runs

Benchmark 2: ./rust/nqueen-new
  Time (mean ± σ):      2.221 s ±  0.013 s    [User: 2.187 s, System: 0.032 s]
  Range (min … max):    2.205 s …  2.242 s    10 runs

Summary
  './rust/nqueen-new' ran
    1.11 ± 0.01 times faster than './rust/nqueen'
```