@i582 2024-01-04
> 2.68 => 2.49 on my laptop

great :D